app/requirements.txt
Django==5.0
djoser==2.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
drf-spectacular==0.26.5
django-cors-headers==4.3.1
gunicorn==22.0.0
psycopg2-binary~=2.0
Pillow==10.1.0
django-environ==0.11.2
openai==1.34.0
python-docx==1.1.2
nltk==3.8
# pywin32==306
